use insta redactions consistently
