+++ crates/symbolicator-service/tests/integration/process_minidump.rs
@@ -6,8 +6,8 @@ use tempfile::NamedTempFile;
 
 use symbolicator_service::types::Scope;
 use symbolicator_test as test;
+use symbolicator_test::assert_snapshot;
 
-use crate::assert_snapshot;
 use crate::symbolication::setup_service;
 
 macro_rules! stackwalk_minidump {
