+++ crates/symbolicator-test/Cargo.toml
@@ -8,10 +8,12 @@ license = "MIT"
 [dependencies]
 axum = "0.6.1"
 humantime = "2.0.1"
+insta = { version = "1.18.0", features = ["redactions", "yaml"] }
+regex = "1.5.5"
 reqwest = { version = "0.11.0", features = ["gzip", "json", "stream", "trust-dns", "multipart"] }
 serde = { version = "1.0.137", features = ["derive", "rc"] }
 symbolicator-sources = { path = "../symbolicator-sources" }
 tempfile = "3.2.0"
-tower-http = { version = "0.3.5", features = ["fs"] }
 tokio = { version = "1.18.1", features = ["rt", "macros", "fs"] }
+tower-http = { version = "0.3.5", features = ["fs"] }
 tracing-subscriber = { version = "0.3.11", features = ["tracing-log", "local-time", "env-filter", "json"] }
